address reviewer comments (minor ones and left some \TODO for mockability)
