+++ software/controller/lib/core/system_constants.h
@@ -44,16 +44,16 @@ static constexpr GPIO::AdcChannel U5ExhaleFlow{GPIO::Port::B, 0, 15};
 
 // Blower is hooked up to PB3 (see [PCB]),
 // which can be linked to Timer2 chanel 2 using AF1 (see [DS], p77)
-static constexpr GPIO::PwmChannel BlowerChannel = {GPIO::Port::B, 3, GPIO::AlternativeFunction::AF1,
-                                                   PeripheralID::Timer2, 2};
+static constexpr GPIO::PwmChannel BlowerChannel{GPIO::Port::B, 3, GPIO::AlternativeFunction::AF1,
+                                                PeripheralID::Timer2, 2};
 // Blower is driven by a 20kHz PWM, as a compromise between resolution and response time
 /// TODO: add/find a better rationale for this, maybe with the resulting response time/resolution
 static constexpr Frequency BlowerFreq{kilohertz(20)};
 
 // psol is hooked up to PA11 (see [PCB]),
 // which can be linked to Timer1 chanel 4 using AF1 (see [DS], p76)
-static constexpr GPIO::PwmChannel PSolChannel = {GPIO::Port::A, 11, GPIO::AlternativeFunction::AF1,
-                                                 PeripheralID::Timer1, 4};
+static constexpr GPIO::PwmChannel PSolChannel{GPIO::Port::A, 11, GPIO::AlternativeFunction::AF1,
+                                              PeripheralID::Timer1, 4};
 // Psol is driven by a 5kHz PWM (\TODO: find rationale behind this?)
 static constexpr Frequency PSolFreq{kilohertz(5)};
 // Testing in Edwin's garage, we found that the psol was fully closed at
@@ -65,8 +65,8 @@ static constexpr float PSolOpen{0.75f};
 
 // Buzzer is hooked up to PB4 (see [PCB]),
 // which can be linked to Timer3 chanel 1 using AF2 (see [DS], p77)
-static constexpr GPIO::PwmChannel BuzzerChannel = {GPIO::Port::B, 4, GPIO::AlternativeFunction::AF2,
-                                                   PeripheralID::Timer3, 1};
+static constexpr GPIO::PwmChannel BuzzerChannel{GPIO::Port::B, 4, GPIO::AlternativeFunction::AF2,
+                                                PeripheralID::Timer3, 1};
 // Buzzer is driven by a 2.4kHz PWM, which produces a medium to high pitch noise, which seems
 // adequate for an alarm of some sort.
 static constexpr Frequency BuzzerFreq{kilohertz(2.4f)};
